TypingTest/pyproject.toml
[tool.poetry]
name = "typingtest"
version = "0.1.0"
description = ""
authors = ["schmaigul <miksuhok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
essential-generators = "^1.0"
invoke = "^1.7.3"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
coverage = "^6.5.0"
pylint = "^2.15.6"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
